+++ media/js/src/containers/dashboard.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useParams, Link } from 'react-router-dom';
-import { Footer } from './footer';
+import { Footer } from '../footer';
 import PropTypes from 'prop-types';
