+++ media/js/src/simulations/simulation1/glossary.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export const GlossaryButton = () => {
+    return (
+        <button
+            className="btn btn-sm btn-primary"
+            data-bs-toggle="modal"
+            data-bs-target="#glossaryModal">
+            Glossary
+        </button>
+    );
+};
