+++ media/js/src/simulations/simulation1/nullHypothesisSection.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import { Katex } from '../katexComponent';
+import { Katex } from '../../utils/katexComponent';
 import PropTypes from 'prop-types';
