+++ media/js/src/simulations/simulation2/components/learningGoals.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { labelIndex } from '../dataAttr';
-import { inlineKatex } from '../utils';
+import { inlineKatex } from '../../../utils/utils';
 
 
 export const LearningGoals = ({
