+++ media/js/src/simulations/simulation2/components/scatterPlot2.jsx
