+++ media/js/src/simulations/simulation2/components/variables.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { inlineKatex, formulaText } from '../utils';
+import { inlineKatex, formulaText } from '../../../utils/utils';
 import { labelIndex as label } from '../dataAttr';
 
 // If this get's implemented into other Simulations, rewrite as a JSX component
