+++ media/js/src/utils/katexComponent.jsx
